fedimint_server_core::dashboard_ui

Trait IDashboardApi

pub trait IDashboardApi {

    // Required methods
    fn auth<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ApiAuth> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn guardian_id<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = PeerId> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn guardian_names<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = BTreeMap<PeerId, String>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn federation_name<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = String> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn session_count<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = u64> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_session_status<'life0, 'async_trait>(
        &'life0 self,
        session_idx: u64,
    ) -> Pin<Box<dyn Future<Output = SessionStatusV2> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn consensus_ord_latency<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Option<Duration>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn p2p_connection_status<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = BTreeMap<PeerId, Option<P2PConnectionStatus>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn federation_invite_code<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = String> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn federation_audit<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = AuditSummary> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn bitcoin_rpc_url<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = SafeUrl> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn bitcoin_rpc_status<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Option<ServerBitcoinRpcStatus>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn download_guardian_config_backup<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        password: &'life1 str,
        guardian_auth: &'life2 GuardianAuthToken,
    ) -> Pin<Box<dyn Future<Output = GuardianConfigBackup> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_module_by_kind(&self, kind: ModuleKind) -> Option<&DynServerModule>;
    fn fedimintd_version<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = String> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn into_dyn(self) -> DynDashboardApi
       where Self: Sized + Send + Sync + 'static { ... }

}

Interface for guardian dashboard API in a running federation

Required Methods

fn auth<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ApiAuth> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the guardian’s authentication details

fn guardian_id<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = PeerId> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the guardian ID

fn guardian_names<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = BTreeMap<PeerId, String>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get a map of peer IDs to guardian names

fn federation_name<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = String> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the federation name

fn session_count<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = u64> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the current active session count

fn get_session_status<'life0, 'async_trait>( &'life0 self, session_idx: u64, ) -> Pin<Box<dyn Future<Output = SessionStatusV2> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get items in a given session

fn consensus_ord_latency<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Option<Duration>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

The time it took to order our last proposal in the current session

fn p2p_connection_status<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = BTreeMap<PeerId, Option<P2PConnectionStatus>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns a map of peer ID to connection status (None = disconnected)

fn federation_invite_code<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = String> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the federation invite code to share with users

fn federation_audit<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = AuditSummary> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the federation audit summary

fn bitcoin_rpc_url<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = SafeUrl> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the url of the bitcoin rpc

fn bitcoin_rpc_status<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Option<ServerBitcoinRpcStatus>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the status of the bitcoin backend

fn download_guardian_config_backup<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, password: &'life1 str, guardian_auth: &'life2 GuardianAuthToken, ) -> Pin<Box<dyn Future<Output = GuardianConfigBackup> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Download a backup of the guardian’s configuration

fn get_module_by_kind(&self, kind: ModuleKind) -> Option<&DynServerModule>

Get reference to a server module instance by module kind

fn fedimintd_version<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = String> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the fedimintd version

Provided Methods

fn into_dyn(self) -> DynDashboardApi
where Self: Sized + Send + Sync + 'static,

Create a trait object